15/09

+++ Curso Python (Guanabara)/Exercicios Python Mundo 1/E23 separado digitos de um numero.py
@@ -1,13 +1,23 @@
-num = int (input('Digite numero'))
+num = float (input('Digite numero'))
+
 unidade = num // 1 % 10  
+
 dezena  = num // 10 % 10 
+
 centena = num // 100 % 10   
+
 milhar = num // 1000 % 10  
 
+
 print (f'analisando o numero {num} ... ')
+
 print (f'unidade: {unidade}') 
+
 print (f'dezena : {dezena}') 
+
 print (f'centena : {centena}') 
+
 print (f'milhar : {milhar}')
+
 
 #ler numero e mostra cada numero separado , como unidade , dezena , centena e milhar
