+++ Curso Python (Guanabara)/Exercicios Python Mundo 1/E28 Jogo advinhação.py
@@ -1,10 +1,16 @@
 from random import randint 
 from time import sleep 
+
 computador = randint (0,5) #faz o computador randomizar
 
 
-player = int (input('Digite um numero'))
+
+player = float (input('Digite um numero'))
+
 print ('Pensando ...')
+
 sleep(3) #pc espera '3' segundos ate executar porxima ação
 
+
 print (f'vc digitou {player}e o computador digitou {computador}')
+
