+++ Curso Python (Guanabara)/Exercicios Python Mundo 1/E30 par ou impar.py
@@ -1,11 +1,21 @@
 
 
-num = int (input('Digite um numero '))
+
+
+num = float (input('Digite um numero '))
+
+
 
 resto = num % 2 
 
+
+
 if resto == 0 : 
+
     print (f'o numero {num} é um numero par ') 
 
+
+
 else : 
+
     print (f'o numero {num} é um numero impar')
