+++ Curso Python (Guanabara)/Exercicios Python Mundo 1/E32 Ano bixesto.py
@@ -1,7 +1,11 @@
 # Descobrindo ano bissexto 
-ano = int (input('qual é o ano para analizar'))
+ano = float (input('qual é o ano para analizar'))
+
 
 if ano % 4 == 0 and ano % 100 != 0 or ano % 400 == 0 : 
+
     print (f'o ano {ano} é bissexto')
+
 else : 
+
     (f'o ano {ano} , nao é bissexto')
