+++ Curso Python (Guanabara)/Exercicios Python Mundo 2/Ex48 Tabuada Usando For.py
@@ -1,5 +1,10 @@
-num = int (input('Digite um numero para a tabuada '))
+num = float (input('Digite um numero para a tabuada '))
+
+
 
 for i in range (0,11): #range usado para numeros 
+
     print (f'{num} x {i} = {i*num}') 
+
 print ('fim')
+
