+++ Curso Python Udemy Luiz Otavio/validado qual valor digitado é maior.py
@@ -1,11 +1,17 @@
-primeiro_valor = int (input('digite um valor '))
-segundo_valor  = int (input('digite outro valor '))
+primeiro_valor = float (input('digite um valor '))
+
+segundo_valor  = float (input('digite outro valor '))
+
 
 
 if segundo_valor > primeiro_valor:
+
     print (f'o segundo valor é {segundo_valor}e é maior que primeiro valor {primeiro_valor} ')
 
+
 elif segundo_valor < primeiro_valor: 
+
     print (f'o segundo valor {segundo_valor} é MENOR que o primeiro valor {primeiro_valor}')
+
