+++ Exercicios_aleatorios/Exercicios Aleatorios/Maior_Menor_numero.py
@@ -1,10 +1,18 @@
 
+
 lista = list()
+
 for num in range (0,5):
-    lista.append (int(input('Digite um numero')))
+
+    lista.append (float(input('Digite um numero')))
+
 
 #for num in enumerate(lista):
+
    #print (f'o numero {num}esta na posiçao{enumerate(lista)}')
 
+
 print ('o menor numero digitado foi ', min(lista))
+
 print ('o maior numero digitado foi ', max(lista))
+
