+++ Exercicios_aleatorios/Exercicios Aleatorios/tabuada.py
@@ -1,4 +1,7 @@
-num=int(input('Tabuada do numero '))
+num=float(input('Tabuada do numero '))
+
+
 
 for n in range (11):
+
     print(f'{num} x {n} = {num*n}')
